requirement.txt
beautifulsoup4==4.13.0b2
google-generativeai==0.5.2
streamlit==1.36.0
